PR-11: Removed the original version of Portfolio
