+++ src/pages/Portfolio/Portfolio.module.less
@@ -1,7 +1,9 @@
 @import "../../Typings/CommonStyles.less";
 
 .projectImage{
-
+    // object-fit: contain;
+    // width: 100%;
+    // height: 100%;
 }
 
 .projectImageBox{
@@ -12,3 +14,15 @@
 .projectImageBox:hover{
     opacity: 0.8;
 }
+.blogTile{
+    background-color: @header-and-content-color;
+}
+
+.blogTitle{
+    color: @bottom-box-info-tile-heading-white;
+}
+
+.readMoreButton{
+    --button-bg: @bottom-box-download-button-orange;
+    --button-color: @bottom-box-info-tile-heading-white;
+}
