--- src/pages/Portfoliov2/Portfolio.module.less
@@ -1,28 +1,0 @@
-@import "../../Typings/CommonStyles.less";
-
-.projectImage{
-    // object-fit: contain;
-    // width: 100%;
-    // height: 100%;
-}
-
-.projectImageBox{
-    background-color: @bottom-box-download-button-orange;
-    opacity: 0;
-    
-}
-.projectImageBox:hover{
-    opacity: 0.8;
-}
-.blogTile{
-    background-color: @header-and-content-color;
-}
-
-.blogTitle{
-    color: @bottom-box-info-tile-heading-white;
-}
-
-.readMoreButton{
-    --button-bg: @bottom-box-download-button-orange;
-    --button-color: @bottom-box-info-tile-heading-white;
-}
